Move Plus Promo modal rendering to a PlusPromoModal shared component. Begin to implement conditional panel display logic
