+++ app/panel/components/Panel.jsx
@@ -13,7 +13,7 @@
 
 import React from 'react';
 import Header from '../containers/HeaderContainer';
-import Modal from '../../shared-components/Modal';
+import { PlusPromoModal } from '../../shared-components';
 import { DynamicUIPortContext } from '../contexts/DynamicUIPortContext';
 import { sendMessage } from '../utils/msg';
 import { setTheme } from '../utils/utils';
@@ -195,7 +195,7 @@ class Panel extends React.Component {
 	}
 
 	_plusPromoClickHandlerPlaceholder = () => {
-		console.error('IVZ Panfel#_plusPromoClickHandlerPlaceholder');
+		console.error('IVZ Panel#_plusPromoClickHandlerPlaceholder');
 	}
 
 	/**
@@ -210,11 +210,15 @@ class Panel extends React.Component {
 
 		const notificationText = this.props.notificationShown && this.renderNotification();
 
+		const { plusPromoModalSeen } = this.props;
+
 		return (
 			<div id="panel">
-				<Modal show>
-					{Modal.renderPlusPromo('inPanel', this._plusPromoClickHandlerPlaceholder)}
-				</Modal>
+				<PlusPromoModal
+					show={!plusPromoModalSeen}
+					location="panel"
+					clickHandler={this._plusPromoClickHandlerPlaceholder}
+				/>
 				<div className="callout-container">
 					<div className={`${(!notificationText ? 'hide ' : '') + this.props.notificationClasses} callout`}>
 						<svg onClick={this.closeNotification} width="15px" height="15px" viewBox="0 0 15 15" className="close-button">
