+++ app/scss/hub.scss
@@ -76,6 +76,7 @@ html, body, #root {
 // Imports from ../shared-components directory
 @import '../shared-components/ExitButton/ExitButton.scss';
 @import '../shared-components/Modal/Modal.scss';
+@import '../shared-components/PlusPromoModal/PlusPromoModal.scss';
 @import '../shared-components/SteppedNavigation/SteppedNavigation.scss';
 @import '../shared-components/ToastMessage/ToastMessage.scss';
 @import '../shared-components/ToggleCheckbox/ToggleCheckbox.scss';
