+++ app/scss/panel.scss
@@ -76,3 +76,4 @@ html body {
 
 // Imports from ../shared-components directory
 @import '../shared-components/Modal/Modal.scss';
+@import '../shared-components/PlusPromoModal/PlusPromoModal.scss';
