+++ app/shared-components/PlusPromoModal/PlusPromoModal.scss
@@ -1,0 +1,200 @@
+.PlusPromoModal__content {
+	position: relative;
+	background-color: #f7f7f7;
+	border: 1.9px solid #930194;
+	z-index: 10;
+
+	&.in-hub {
+		width: 646px;
+		height: 553px;
+	}
+
+	&.in-panel {
+		width: 556px;
+		height: 471px;
+	}
+}
+
+.PlusPromoModal__buttons-background {
+	background-color: #e7ecee;
+	position: absolute;
+	bottom: 3px;
+	width: 99%;
+	height: 72px;
+	z-index: -1;
+}
+
+.PlusPromoModal__thanks-for-download {
+	padding-top: 30px;
+	font-size: 26px;
+	font-weight: bold;
+}
+
+.PlusPromoModal__choose-your-plan {
+	font-size: 18px;
+	font-weight: bold;
+
+	&.in-hub { padding-top: 10px; }
+	&.in-panel { padding-top: 17px; }
+}
+
+.PlusPromoModal__options-container {
+	height: 100%;
+	margin-top: 44px;
+	display: flex;
+
+	&.in-hub { justify-content: space-evenly; }
+	&.in-panel { justify-content: space-around; }
+}
+
+.PlusPromoModal__option-container {
+	display: flex;
+	align-items: center;
+	flex-direction: column;
+}
+
+.PlusPromoModal__option-description-box {
+	border-style: solid;
+	border-width: 3px;
+	background-color: #FFFFFF;
+
+	&.basic {
+		border-color: #1dafed;
+	}
+
+	&.plus {
+		position: relative; // so that the recommended banner is positioned relative to this element
+		border-image-source: linear-gradient(39deg, #f2daa2, #eab968);
+		border-image-slice: 1;
+	}
+
+	&.in-hub {
+		margin-bottom: 45px;
+		width: 240px;
+		height: 298px;
+	}
+
+	&.in-panel {
+		margin-bottom: 30px;
+		width: 221px;
+		height: 294px;
+	}
+}
+
+.PlusPromoModal__recommended-banner {
+	position: absolute;
+
+	&.in-hub {
+		left: -30px;
+		top: -30px;
+	}
+
+	&.in-panel {
+		left: -10px;
+		top: -32px;
+	}
+}
+
+.PlusPromoModal__recommended-banner-text {
+	position: relative;
+	left: 18px;
+	top: -47px;
+	transform: rotate(-15deg);
+	color: white;
+	font-size: 16px;
+	font-weight: bold;
+}
+
+.PlusPromoModal__option-header {
+	margin-top: 20px;
+	font-size: 19px;
+	font-weight: bold;
+	letter-spacing: -0.5px;
+	text-align: center;
+
+	&.basic { color: #1dafed; }
+	&.plus { color: #ebbf73; }
+}
+
+.PlusPromoModal__price-text {
+	line-height: 1.3;
+	text-align: center;
+
+	&.basic { color: #1dafed; }
+	&.plus { color: #ebbf73; }
+}
+
+.PlusPromoModal__currency-sign {
+	font-size: 30px;
+	font-weight: bold;
+}
+
+.PlusPromoModal__amount {
+	vertical-align: middle;
+	font-family: "Open Sans", "Helvetica Neue", Helvetica, Arial, sans-serif;
+	font-size: 58px;
+}
+
+.PlusPromoModal__per-month {
+	vertical-align: sub;
+	font-size: 18px;
+	font-weight: bold;
+}
+
+.PlusPromoModal__option-description-item {
+	padding-bottom: 15px;
+	font-size: 16px;
+	text-align: center;
+	color: #333333;
+
+	&.italic {
+		font-style: italic;
+	}
+}
+
+.PlusPromoModal__check-icon {
+	padding-right: 5px;
+	margin-top: -18px;
+}
+
+.PlusPromoModal__plus-option-description-item-container {
+	display: flex;
+	justify-content: center;
+}
+
+.PlusPromoModal__button {
+	margin: 0;
+	font-size: 11px;
+	font-weight: 600;
+	text-transform: uppercase;
+	letter-spacing: 0.5px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	box-shadow: none;
+
+	&.basic {
+		width: 135px;
+		height: 40px;
+		border: solid 2px #15b4f2;
+		background-color: white;
+		color: #2cbcf4;
+	}
+	&.basic:hover {
+		background-color: #2cbcf4;
+		color: white;
+	}
+
+	&.plus {
+		width: 163px;
+		height: 38px;
+		border: none;
+		background-image: linear-gradient(to bottom, #2fdbfa, #15b4f2);
+		color: white;
+		font-weight: 600;
+		letter-spacing: 0.5px;
+	}
+	&.plus:hover {
+		background-image: linear-gradient(to bottom, #1fcbea, #05a4e2);
+	}
+}
