+++ app/shared-components/index.js
@@ -17,6 +17,7 @@
 
 import ExitButton from './ExitButton';
 import Modal from './Modal';
+import PlusPromoModal from './PlusPromoModal';
 import SteppedNavigation from './SteppedNavigation';
 import ToastMessage from './ToastMessage';
 import ToggleCheckbox from './ToggleCheckbox';
@@ -25,6 +26,7 @@ import ToggleSwitch from './ToggleSwitch';
 export {
 	ExitButton,
 	Modal,
+	PlusPromoModal,
 	SteppedNavigation,
 	ToastMessage,
 	ToggleCheckbox,
