+++ src/classes/ConfData.js
@@ -123,6 +123,7 @@ class ConfData {
 			_initProperty('notify_library_updates', false);
 			_initProperty('notify_upgrade_updates', true);
 			_initProperty('paid_subscription', false);
+			_initProperty('plus_promo_modal_last_seen', null);
 			_initProperty('rewards_accepted', false);
 			_initProperty('rewards_opted_in', false);
 			_initProperty('settings_last_imported', 0);
