+++ src/classes/PanelData.js
@@ -351,7 +351,7 @@ class PanelData {
 		const { id: tab_id } = this._activeTab;
 		const {
 			current_theme, enable_ad_block, enable_anti_tracking, enable_smart_block,
-			enable_offers, is_expanded, is_expert, language, reload_banner_status,
+			enable_offers, is_expanded, is_expert, language, plus_promo_modal_last_seen, reload_banner_status,
 			trackers_banner_status,
 		} = conf;
 
@@ -365,6 +365,7 @@ class PanelData {
 			is_expert,
 			is_android: globals.BROWSER_INFO.os === 'android',
 			language,
+			plusPromoModalSeen: !!plus_promo_modal_last_seen,
 			reload_banner_status,
 			tab_id,
 			trackers_banner_status,
